shared: consolidate all Tcl list + escape/unescape manipulation
Collapse the ~13 scattered, subtly-divergent re-implementations of Tcl list
semantics into one canonical home so every concern shares the same rules
instead of re-deriving brace/quote/backslash handling.

shared/tcl_list.py (renamed from tcl_quoting.py) now owns:
  - tcl_list_quote  (TclScanElement/TclConvertElement port — already canonical)
  - tcl_list_split  (Tcl_SplitList: brace/quote/backslash-aware; strict mode
                     raises, optional unescape recovers element values)
  - tcl_list_join   (inverse of split — what Tcl's `list` builtin produces)

shared/tcl_subst.py (moved from compiler/parsing/substitution.py) owns the
backslash-substitution primitive (escape decoding); callers now import it from
shared, and tcl_list_split uses it for value-recovering splits.

Migrated every duplicate to the canonical functions:
  - quoting: wasm _core info-args spec, bytecode _helpers/_list/_dict/_cmd_subst
    /_values, dialects const_fold, minifier, vm _list_escape, regexp/test-support
    vm commands, irule_test bridge.
  - splitting: tcl_expr_eval, bytecode _helpers, vm _split_list, regexp/minifier.
  - backslash_subst: ~15 importers repointed to shared.tcl_subst.

Because the canonical quoter/splitter match C Tcl 9 exactly (verified against
tclsh's `list` builtin and round-trips), the VM now produces correct output for
30+ cases previously listed as known failures — removed from the
list/dict/proc/string known-failure sets.  const_fold's only compiler import is
gone, so its import-linter carve-out is dropped too.
