+++ analyser/_analyser/_scope.py
@@ -8,7 +8,6 @@ if TYPE_CHECKING:
 else:
     _Base = object
 
-from compiler.parsing.substitution import backslash_subst as _backslash_subst
 from shared.naming import (
     is_brace_substitutable,
 )
@@ -22,6 +21,7 @@ from shared.naming import (
     split_array_name as _split_array_name,
 )
 from shared.ranges import position_from_relative, range_from_token
+from shared.tcl_subst import backslash_subst as _backslash_subst
 from shared.tokens import Token, TokenType
 
 from ..semantic_model import (
